Hyunjun/feat/jwt httponly & social login

+++ src/main/java/MATE/Carpool/config/jwt/RefreshToken.java
@@ -19,5 +19,5 @@ public class RefreshToken {
     @Column(unique=true,nullable = false)
     private String memberId;
     @Column
-    private Long expiresAt;
+    private Integer expiresAt;
 }
